+++ Derivatives/LibreFace.App.Consoles/State.cs
@@ -1,0 +1,9 @@
+﻿namespace LibreFace.App.Consoles {
+    internal enum State {
+        Unspecified,
+        Pending,
+        Processing,
+        Done,
+        Error,
+    }
+}