+++ Derivatives/LibreFace.App.Consoles/RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+# 1.2.0
+
+Optimized CPU performance.
+Enabled parallel processing.
+Added more options.
+Updated UI.
+
 # 1.1.0
 
 Update LibreFace package to 2.0.0.